+++ R/ollama.R
@@ -58,6 +58,7 @@ create_request <- function(endpoint, host = NULL) {
 #' @param prompt A character string of the prompt like "The sky is..."
 #' @param suffix A character string after the model response. Default is "".
 #' @param images A path to an image file to include in the prompt. Default is "".
+#' @param format Format to return a response in. Format can be json/list (structured response).
 #' @param system A character string of the system prompt (overrides what is defined in the Modelfile). Default is "".
 #' @param template A character string of the prompt template (overrides what is defined in the Modelfile). Default is "".
 #' @param context A list of context from a previous response to include previous conversation in the prompt. Default is an empty list.
@@ -86,10 +87,10 @@ create_request <- function(endpoint, host = NULL) {
 #' image_path <- file.path(system.file("extdata", package = "ollamar"), "image1.png")
 #' # use vision or multimodal model such as https://ollama.com/benzie/llava-phi-3
 #' generate("benzie/llava-phi-3:latest", "What is in the image?", images = image_path, output = "text")
-generate <- function(model, prompt, suffix = "", images = "", system = "", template = "", context = list(), stream = FALSE, raw = FALSE, keep_alive = "5m", output = c("resp", "jsonlist", "raw", "df", "text", "req"), endpoint = "/api/generate", host = NULL, ...) {
+generate <- function(model, prompt, suffix = "", images = "", format = list(), system = "", template = "", context = list(), stream = FALSE, raw = FALSE, keep_alive = "5m", output = c("resp", "jsonlist", "raw", "df", "text", "req", "structured"), endpoint = "/api/generate", host = NULL, ...) {
     output <- output[1]
-    if (!output %in% c("df", "resp", "jsonlist", "raw", "text", "req")) {
-        stop("Invalid output format specified. Supported formats: 'df', 'resp', 'jsonlist', 'raw', 'text', 'req'")
+    if (!output %in% c("df", "resp", "jsonlist", "raw", "text", "req", "structured")) {
+        stop("Invalid output format specified. Supported formats: 'df', 'resp', 'jsonlist', 'raw', 'text', 'req', 'structured'")
     }
 
     req <- create_request(endpoint, host)
@@ -111,6 +112,10 @@ generate <- function(model, prompt, suffix = "", images = "", system = "", templ
         stream = stream,
         keep_alive = keep_alive
     )
+
+    if (length(format) != 0 & inherits(format, "list")) {
+        body_json$format <- format
+    }
 
     # check if model options are passed and specified correctly
     opts <- list(...)
@@ -169,8 +174,9 @@ generate <- function(model, prompt, suffix = "", images = "", system = "", templ
 #' @param messages A list with list of messages for the model (see examples below).
 #' @param tools Tools for the model to use if supported. Requires stream = FALSE. Default is an empty list.
 #' @param stream Enable response streaming. Default is FALSE.
+#' @param format Format to return a response in. Format can be json/list (structured response).
 #' @param keep_alive The duration to keep the connection alive. Default is "5m".
-#' @param output The output format. Default is "resp". Other options are "jsonlist", "raw", "df", "text", "req" (httr2_request object), "tools" (tool calling)
+#' @param output The output format. Default is "resp". Other options are "jsonlist", "raw", "df", "text", "req" (httr2_request object), "tools" (tool calling), "structured" (structured output)
 #' @param endpoint The endpoint to chat with the model. Default is "/api/chat".
 #' @param host The base URL to use. Default is NULL, which uses Ollama's default base URL.
 #' @param ... Additional options to pass to the model.
@@ -208,10 +214,10 @@ generate <- function(model, prompt, suffix = "", images = "", system = "", templ
 #'    list(role = "user", content = "What is in the image?", images = image_path)
 #' )
 #' chat("benzie/llava-phi-3", messages, output = 'text')
-chat <- function(model, messages, tools = list(), stream = FALSE, keep_alive = "5m", output = c("resp", "jsonlist", "raw", "df", "text", "req", "tools"), endpoint = "/api/chat", host = NULL, ...) {
+chat <- function(model, messages, tools = list(), stream = FALSE, format = list(), keep_alive = "5m", output = c("resp", "jsonlist", "raw", "df", "text", "req", "tools", "structured"), endpoint = "/api/chat", host = NULL, ...) {
     output <- output[1]
-    if (!output %in% c("df", "resp", "jsonlist", "raw", "text", "req", "tools")) {
-        stop("Invalid output format specified. Supported formats: 'df', 'resp', 'jsonlist', 'raw', 'text'")
+    if (!output %in% c("df", "resp", "jsonlist", "raw", "text", "req", "tools", "structured")) {
+        stop("Invalid output format specified. Supported formats: 'df', 'resp', 'jsonlist', 'raw', 'text', 'tools', 'structured'")
     }
 
     req <- create_request(endpoint, host)
@@ -230,6 +236,10 @@ chat <- function(model, messages, tools = list(), stream = FALSE, keep_alive = "
         stream = stream,
         keep_alive = keep_alive
     )
+
+    if (length(format) != 0 & inherits(format, "list")) {
+        body_json$format <- format
+    }
 
     opts <- list(...)
     if (length(opts) > 0) {